pyproject.toml
[project]
name = "itsg33-accreditation"
version = "0.1.0"
description = "ITSG-33 Accreditation System using Multi-Agent AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "swarms>=6.0.0",
    "google-generativeai>=0.8.0",
    "fastmcp>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
    "pillow>=11.0.0",
    "chromadb>=0.5.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
